pyproject.toml
[project]
name = "azure-templates"
version = "0.1.0"
description = ""
authors = [
    {name = "Tammy DiPrima",email = "tammy.diprima@stonybrook.edu"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "azure-identity (>=1.24.0,<2.0.0)",
    "azure-mgmt-resource (>=24.0.0,<25.0.0)",
    "azure-mgmt-compute (>=36.0.0,<37.0.0)",
    "azure-mgmt-network (>=29.0.0,<30.0.0)",
    "azure-mgmt-storage (>=23.0.1,<24.0.0)",
    "azure-storage-blob (>=12.26.0,<13.0.0)",
    "azure-ai-ml (>=1.28.1,<2.0.0)",
    "openai (>=1.102.0,<2.0.0)",
    "azure-search-documents (>=11.5.3,<12.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]

[tool.poetry]
packages = [{include = "azure_templates", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
